Separate ClosestNotMeRayResultCallback
